refactor(api): deduplicate shared controller request schemas into controllers/common/
Extract duplicate Pydantic request schemas that were copy-pasted across web, console/explore, and service_api controllers into shared modules under controllers/common/:

- ConversationRenamePayload (3 copies → 1)
- MessageListQuery + MessageFeedbackPayload (3 copies each → 1)
- SavedMessageListQuery + SavedMessageCreatePayload (2 copies each → 1)
- WorkflowRunPayload (3 copies → 1, service_api extends with response_mode)
- TextToAudioPayload (2 copies → 1)
