+++ api/controllers/common/saved_message_schemas.py
@@ -1,0 +1,12 @@
+from pydantic import BaseModel, Field
+
+from libs.helper import UUIDStrOrEmpty
+
+
+class SavedMessageListQuery(BaseModel):
+    last_id: UUIDStrOrEmpty | None = None
+    limit: int = Field(default=20, ge=1, le=100)
+
+
+class SavedMessageCreatePayload(BaseModel):
+    message_id: UUIDStrOrEmpty
