+++ api/controllers/common/workflow_schemas.py
@@ -1,0 +1,8 @@
+from typing import Any
+
+from pydantic import BaseModel
+
+
+class WorkflowRunPayload(BaseModel):
+    inputs: dict[str, Any]
+    files: list[dict[str, Any]] | None = None
