+++ api/controllers/web/audio.py
@@ -3,10 +3,10 @@ import logging
 from flask import request
 from flask_restx import fields, marshal_with
 from graphon.model_runtime.errors.invoke import InvokeError
-from pydantic import BaseModel, field_validator
 from werkzeug.exceptions import InternalServerError
 
 import services
+from controllers.common.audio_schemas import TextToAudioPayload
 from controllers.web import web_ns
 from controllers.web.error import (
     AppUnavailableError,
@@ -21,7 +21,6 @@ from controllers.web.error import (
 )
 from controllers.web.wraps import WebApiResource
 from core.errors.error import ModelCurrentlyNotSupportError, ProviderTokenNotInitError, QuotaExceededError
-from libs.helper import uuid_value
 from models.model import App
 from services.audio_service import AudioService
 from services.errors.audio import (
@@ -32,21 +31,6 @@ from services.errors.audio import (
 )
 
 from ..common.schema import register_schema_models
-
-
-class TextToAudioPayload(BaseModel):
-    message_id: str | None = None
-    voice: str | None = None
-    text: str | None = None
-    streaming: bool | None = None
-
-    @field_validator("message_id")
-    @classmethod
-    def validate_message_id(cls, value: str | None) -> str | None:
-        if value is None:
-            return value
-        return uuid_value(value)
-
 
 register_schema_models(web_ns, TextToAudioPayload)
